requirements.txt
Django==2.0.5
psycopg2==2.7.*
dj-database-url==0.4.2
djangorestframework==3.8.2